fix: lets apply the build pack before the cluster PR
so that we can properly determine what kind of scheduler we need to use in the cluster git repository PR.

Then we do the commit + push after the PR is merged and the webhooks + triggers are setup
